refactor(password-reset): move to account mail controller
