+++ infrastructure/src/main/java/com/kaua/ecommerce/users/infrastructure/api/AccountAPI.java
@@ -1,11 +1,11 @@
 package com.kaua.ecommerce.users.infrastructure.api;
 
 import com.kaua.ecommerce.users.infrastructure.accounts.models.CreateAccountApiInput;
-import com.kaua.ecommerce.users.infrastructure.accounts.models.RequestResetPasswordApiInput;
-import com.kaua.ecommerce.users.infrastructure.accounts.models.ResetPasswordApiInput;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @RequestMapping(value = "accounts")
 public interface AccountAPI {
@@ -15,10 +15,4 @@ public interface AccountAPI {
             produces = MediaType.APPLICATION_JSON_VALUE
     )
     ResponseEntity<?> createAccount(@RequestBody CreateAccountApiInput input);
-
-    @PostMapping("/request-reset-password")
-    ResponseEntity<?> requestResetPassword(@RequestBody RequestResetPasswordApiInput input);
-
-    @PatchMapping("/reset-password/{token}")
-    ResponseEntity<?> resetPassword(@RequestBody ResetPasswordApiInput input, @PathVariable String token);
 }
