+++ infrastructure/src/main/java/com/kaua/ecommerce/users/infrastructure/api/controllers/AccountMailController.java
@@ -4,9 +4,15 @@ import com.kaua.ecommerce.users.application.account.mail.confirm.ConfirmAccountM
 import com.kaua.ecommerce.users.application.account.mail.confirm.ConfirmAccountMailUseCase;
 import com.kaua.ecommerce.users.application.account.mail.create.CreateAccountMailCommand;
 import com.kaua.ecommerce.users.application.account.mail.create.CreateAccountMailUseCase;
+import com.kaua.ecommerce.users.application.account.update.password.RequestResetPasswordCommand;
+import com.kaua.ecommerce.users.application.account.update.password.RequestResetPasswordUseCase;
+import com.kaua.ecommerce.users.application.account.update.password.reset.ResetPasswordCommand;
+import com.kaua.ecommerce.users.application.account.update.password.reset.ResetPasswordUseCase;
 import com.kaua.ecommerce.users.domain.accounts.mail.AccountMailType;
 import com.kaua.ecommerce.users.domain.utils.IdUtils;
 import com.kaua.ecommerce.users.domain.utils.InstantUtils;
+import com.kaua.ecommerce.users.infrastructure.accounts.models.RequestResetPasswordApiInput;
+import com.kaua.ecommerce.users.infrastructure.accounts.models.ResetPasswordApiInput;
 import com.kaua.ecommerce.users.infrastructure.api.AccountMailAPI;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,13 +25,19 @@ public class AccountMailController implements AccountMailAPI {
 
     private final CreateAccountMailUseCase createAccountMailUseCase;
     private final ConfirmAccountMailUseCase confirmAccountMailUseCase;
+    private final RequestResetPasswordUseCase requestResetPasswordUseCase;
+    private final ResetPasswordUseCase resetPasswordUseCase;
 
     public AccountMailController(
             final CreateAccountMailUseCase createAccountMailUseCase,
-            final ConfirmAccountMailUseCase confirmAccountMailUseCase
+            final ConfirmAccountMailUseCase confirmAccountMailUseCase,
+            final RequestResetPasswordUseCase requestResetPasswordUseCase,
+            final ResetPasswordUseCase resetPasswordUseCase
     ) {
         this.createAccountMailUseCase = createAccountMailUseCase;
         this.confirmAccountMailUseCase = confirmAccountMailUseCase;
+        this.requestResetPasswordUseCase = requestResetPasswordUseCase;
+        this.resetPasswordUseCase = resetPasswordUseCase;
     }
 
     @Override
@@ -53,5 +65,21 @@ public class AccountMailController implements AccountMailAPI {
         return aResult.isLeft()
                 ? ResponseEntity.unprocessableEntity().body(aResult.getLeft())
                 : ResponseEntity.status(HttpStatus.CREATED).body(aResult.getRight());
+    }
+
+    @Override
+    public ResponseEntity<?> requestResetPassword(RequestResetPasswordApiInput input) {
+        this.requestResetPasswordUseCase.execute(RequestResetPasswordCommand.with(input.email()));
+        return ResponseEntity.noContent().build();
+    }
+
+    @Override
+    public ResponseEntity<?> resetPassword(ResetPasswordApiInput input, String token) {
+        final var aResult = this.resetPasswordUseCase.execute(ResetPasswordCommand
+                .with(token, input.password()));
+
+        return aResult.isLeft()
+                ? ResponseEntity.unprocessableEntity().body(aResult.getLeft())
+                : ResponseEntity.noContent().build();
     }
 }
