+++ infrastructure/src/test/java/com/kaua/ecommerce/users/infrastructure/api/AccountAPITest.java
@@ -5,24 +5,18 @@ import com.kaua.ecommerce.users.ControllerTest;
 import com.kaua.ecommerce.users.application.account.create.CreateAccountCommand;
 import com.kaua.ecommerce.users.application.account.create.CreateAccountOutput;
 import com.kaua.ecommerce.users.application.account.create.CreateAccountUseCase;
-import com.kaua.ecommerce.users.application.account.mail.confirm.ConfirmAccountMailCommand;
 import com.kaua.ecommerce.users.application.account.update.password.RequestResetPasswordUseCase;
-import com.kaua.ecommerce.users.application.account.update.password.reset.ResetPasswordCommand;
 import com.kaua.ecommerce.users.application.account.update.password.reset.ResetPasswordUseCase;
 import com.kaua.ecommerce.users.application.either.Either;
-import com.kaua.ecommerce.users.domain.exceptions.NotFoundException;
 import com.kaua.ecommerce.users.domain.utils.RandomStringUtils;
 import com.kaua.ecommerce.users.domain.validation.Error;
 import com.kaua.ecommerce.users.domain.validation.handler.NotificationHandler;
 import com.kaua.ecommerce.users.infrastructure.accounts.models.CreateAccountApiInput;
-import com.kaua.ecommerce.users.infrastructure.accounts.models.RequestResetPasswordApiInput;
-import com.kaua.ecommerce.users.infrastructure.accounts.models.ResetPasswordApiInput;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
@@ -51,9 +45,6 @@ public class AccountAPITest {
 
     @MockBean
     private ResetPasswordUseCase resetPasswordUseCase;
-
-    @MockBean
-    private BCryptPasswordEncoder passwordEncoder;
 
     @Test
     public void givenAValidCommand_whenCallCreateAccount_thenShouldReturneAnAccountId() throws Exception {
@@ -440,93 +431,5 @@ public class AccountAPITest {
                         Objects.equals(aEmail, cmd.email()) &&
                         Objects.equals(aPassword, cmd.password())
         ));
-    }
-
-    @Test
-    public void givenAValidCommand_whenCallRequestResetPassword_thenShouldReturnNotContent() throws Exception {
-        // given
-        final var aEmail = "teste@teste.com";
-
-        final var aInput = new RequestResetPasswordApiInput(aEmail);
-
-        final var request = MockMvcRequestBuilders.post("/accounts/request-reset-password")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(this.mapper.writeValueAsString(aInput));
-
-        this.mvc.perform(request)
-                .andDo(MockMvcResultHandlers.print())
-                .andExpect(MockMvcResultMatchers.status().isNoContent());
-
-        Mockito.verify(requestResetPasswordUseCase, Mockito.times(1)).execute(argThat(cmd ->
-                Objects.equals(aEmail, cmd.email())
-        ));
-    }
-
-    @Test
-    public void givenAInvalidCommandEmailNotFound_whenCallRequestResetPassword_thenShouldThrowNotFoundException() throws Exception {
-        // given
-        final var aEmail = "teste@teste.com";
-
-        final var aInput = new RequestResetPasswordApiInput(aEmail);
-
-        Mockito.doThrow(NotFoundException.class)
-                .when(requestResetPasswordUseCase).execute(Mockito.any());
-
-        final var request = MockMvcRequestBuilders.post("/accounts/request-reset-password")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(this.mapper.writeValueAsString(aInput));
-
-        this.mvc.perform(request)
-                .andDo(MockMvcResultHandlers.print())
-                .andExpect(MockMvcResultMatchers.status().isNotFound());
-
-        Mockito.verify(requestResetPasswordUseCase, Mockito.times(1)).execute(argThat(cmd ->
-                Objects.equals(aEmail, cmd.email())
-        ));
-    }
-
-    @Test
-    public void givenAValidCommand_whenCallResetPassword_thenShouldReturneNoContent() throws Exception {
-        // given
-        final var aToken = RandomStringUtils.generateValue(36);
-        final var aPassword = "1234567Ab*";
-
-        final var aInput = new ResetPasswordApiInput(aPassword);
-
-        Mockito.when(resetPasswordUseCase.execute(Mockito.any(ResetPasswordCommand.class)))
-                .thenReturn(Either.right(true));
-
-        final var request = MockMvcRequestBuilders.patch("/accounts/reset-password/{token}", aToken)
-                .accept(MediaType.APPLICATION_JSON)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(this.mapper.writeValueAsString(aInput));
-
-        this.mvc.perform(request)
-                .andDo(MockMvcResultHandlers.print())
-                .andExpect(MockMvcResultMatchers.status().isNoContent());
-    }
-
-    @Test
-    public void givenAnInvalidTokenExpired_whenCallResetPassword_thenShouldReturneAnError() throws Exception {
-        // given
-        final var expectedErrorMessage = "Token expired";
-        final var aToken = RandomStringUtils.generateValue(36);
-        final var aPassword = "1234567Ab*";
-
-        final var aInput = new ResetPasswordApiInput(aPassword);
-
-        Mockito.when(resetPasswordUseCase.execute(Mockito.any(ResetPasswordCommand.class)))
-                .thenReturn(Either.left(NotificationHandler.create(new Error(expectedErrorMessage))));
-
-        final var request = MockMvcRequestBuilders.patch("/accounts/reset-password/{token}", aToken)
-                .accept(MediaType.APPLICATION_JSON)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(this.mapper.writeValueAsString(aInput));
-
-        this.mvc.perform(request)
-                .andDo(MockMvcResultHandlers.print())
-                .andExpect(MockMvcResultMatchers.status().isUnprocessableEntity())
-                .andExpect(MockMvcResultMatchers.jsonPath("errors", hasSize(1)))
-                .andExpect(MockMvcResultMatchers.jsonPath("$.errors[0].message", equalTo(expectedErrorMessage)));
     }
 }
