+++ infrastructure/src/test/java/com/kaua/ecommerce/users/infrastructure/api/AccountMailAPITest.java
@@ -6,6 +6,9 @@ import com.kaua.ecommerce.users.application.account.mail.confirm.ConfirmAccountM
 import com.kaua.ecommerce.users.application.account.mail.confirm.ConfirmAccountMailUseCase;
 import com.kaua.ecommerce.users.application.account.mail.create.CreateAccountMailOutput;
 import com.kaua.ecommerce.users.application.account.mail.create.CreateAccountMailUseCase;
+import com.kaua.ecommerce.users.application.account.update.password.RequestResetPasswordUseCase;
+import com.kaua.ecommerce.users.application.account.update.password.reset.ResetPasswordCommand;
+import com.kaua.ecommerce.users.application.account.update.password.reset.ResetPasswordUseCase;
 import com.kaua.ecommerce.users.application.either.Either;
 import com.kaua.ecommerce.users.domain.accounts.Account;
 import com.kaua.ecommerce.users.domain.exceptions.DomainException;
@@ -13,6 +16,8 @@ import com.kaua.ecommerce.users.domain.exceptions.NotFoundException;
 import com.kaua.ecommerce.users.domain.utils.RandomStringUtils;
 import com.kaua.ecommerce.users.domain.validation.Error;
 import com.kaua.ecommerce.users.domain.validation.handler.NotificationHandler;
+import com.kaua.ecommerce.users.infrastructure.accounts.models.RequestResetPasswordApiInput;
+import com.kaua.ecommerce.users.infrastructure.accounts.models.ResetPasswordApiInput;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,8 +28,11 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
+import java.util.Objects;
+
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasSize;
+import static org.mockito.ArgumentMatchers.argThat;
 
 @ControllerTest(controllers = AccountMailAPI.class)
 public class AccountMailAPITest {
@@ -40,6 +48,12 @@ public class AccountMailAPITest {
 
     @MockBean
     private ConfirmAccountMailUseCase confirmAccountMailUseCase;
+
+    @MockBean
+    private RequestResetPasswordUseCase requestResetPasswordUseCase;
+
+    @MockBean
+    private ResetPasswordUseCase resetPasswordUseCase;
 
     @Test
     public void givenAValidCommand_whenCallConfirmAccount_thenShouldReturneNoContent() throws Exception {
@@ -254,5 +268,93 @@ public class AccountMailAPITest {
                 .andExpect(MockMvcResultMatchers.jsonPath("$.message", equalTo(expectedErrorMessage)));
 
         Mockito.verify(createAccountMailUseCase, Mockito.times(1)).execute(Mockito.any());
+    }
+
+    @Test
+    public void givenAValidCommand_whenCallRequestResetPassword_thenShouldReturnNotContent() throws Exception {
+        // given
+        final var aEmail = "teste@teste.com";
+
+        final var aInput = new RequestResetPasswordApiInput(aEmail);
+
+        final var request = MockMvcRequestBuilders.post("/accounts/request-reset-password")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(this.mapper.writeValueAsString(aInput));
+
+        this.mvc.perform(request)
+                .andDo(MockMvcResultHandlers.print())
+                .andExpect(MockMvcResultMatchers.status().isNoContent());
+
+        Mockito.verify(requestResetPasswordUseCase, Mockito.times(1)).execute(argThat(cmd ->
+                Objects.equals(aEmail, cmd.email())
+        ));
+    }
+
+    @Test
+    public void givenAInvalidCommandEmailNotFound_whenCallRequestResetPassword_thenShouldThrowNotFoundException() throws Exception {
+        // given
+        final var aEmail = "teste@teste.com";
+
+        final var aInput = new RequestResetPasswordApiInput(aEmail);
+
+        Mockito.doThrow(NotFoundException.class)
+                .when(requestResetPasswordUseCase).execute(Mockito.any());
+
+        final var request = MockMvcRequestBuilders.post("/accounts/request-reset-password")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(this.mapper.writeValueAsString(aInput));
+
+        this.mvc.perform(request)
+                .andDo(MockMvcResultHandlers.print())
+                .andExpect(MockMvcResultMatchers.status().isNotFound());
+
+        Mockito.verify(requestResetPasswordUseCase, Mockito.times(1)).execute(argThat(cmd ->
+                Objects.equals(aEmail, cmd.email())
+        ));
+    }
+
+    @Test
+    public void givenAValidCommand_whenCallResetPassword_thenShouldReturneNoContent() throws Exception {
+        // given
+        final var aToken = RandomStringUtils.generateValue(36);
+        final var aPassword = "1234567Ab*";
+
+        final var aInput = new ResetPasswordApiInput(aPassword);
+
+        Mockito.when(resetPasswordUseCase.execute(Mockito.any(ResetPasswordCommand.class)))
+                .thenReturn(Either.right(true));
+
+        final var request = MockMvcRequestBuilders.patch("/accounts/reset-password/{token}", aToken)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(this.mapper.writeValueAsString(aInput));
+
+        this.mvc.perform(request)
+                .andDo(MockMvcResultHandlers.print())
+                .andExpect(MockMvcResultMatchers.status().isNoContent());
+    }
+
+    @Test
+    public void givenAnInvalidTokenExpired_whenCallResetPassword_thenShouldReturneAnError() throws Exception {
+        // given
+        final var expectedErrorMessage = "Token expired";
+        final var aToken = RandomStringUtils.generateValue(36);
+        final var aPassword = "1234567Ab*";
+
+        final var aInput = new ResetPasswordApiInput(aPassword);
+
+        Mockito.when(resetPasswordUseCase.execute(Mockito.any(ResetPasswordCommand.class)))
+                .thenReturn(Either.left(NotificationHandler.create(new Error(expectedErrorMessage))));
+
+        final var request = MockMvcRequestBuilders.patch("/accounts/reset-password/{token}", aToken)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(this.mapper.writeValueAsString(aInput));
+
+        this.mvc.perform(request)
+                .andDo(MockMvcResultHandlers.print())
+                .andExpect(MockMvcResultMatchers.status().isUnprocessableEntity())
+                .andExpect(MockMvcResultMatchers.jsonPath("errors", hasSize(1)))
+                .andExpect(MockMvcResultMatchers.jsonPath("$.errors[0].message", equalTo(expectedErrorMessage)));
     }
 }
